requirements.txt
discord.py>=2.0.0
line-bot-sdk
flask
requests
pydub
apnggif